Faster execution of Galaxy.__init__
Also — turns out I overwrote the “(planets_)discovered_by” attributes
with lists when they would normally be sets. That was why the “add”
method wouldn’t work, requiring the “append” method. Oops.

As of the last commit (#93), the code on PythonAnywhere is no longer
returning 500 errors but is instead timing out. So that’s progress. I
guess.